requirements.txt
aiosqlite==0.18.0
alembic==1.9.3
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.1.0
fastapi==0.91.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
pydantic==1.10.4
PyJWT==2.6.0
pytest==7.2.1
pytest-asyncio==0.20.3
python-dotenv==0.21.1
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
starlette==0.24.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
